Creates new Package for sprite pipeline.
New Package engine.sprites will handle the loading and management
of sprite assets. Currently consists of 4 classes each representing
a different aspect of the assets.

Sprite maintains the data of the entire asset, including listings
of Frames and Animations as well as the asset name and a url to the
source image.

Animation contains information used for animating a Sprite, the
frame sequence and whether or not to repeat the animation once
completed.

Frame merely holds the location on the source image for a single
frame of the Sprite.

SpriteReference is for tracking instances of the sprite in the
engine. Keeps the name, current frame, and animation name.